Changed strnum to use int64_t instead of double, as none of this application using floating point math, just lots of int64_t.  Much less casting all over the place, can't hurt performance.
Consolidated all DB setup into new function setup_nldb(db), cleaned up cmd line app and unit tests.

Added nodes::get_path(node), can be used to quickly see if a node is contained by another.

Search is full coded and lightly tested.

+++ src/nldblib/dbreader.cpp
@@ -52,9 +52,9 @@ namespace nldb
         switch (columnType)
         {
         case SQLITE_INTEGER:
-            return static_cast<double>(sqlite3_column_int64(m_stmt, idx));
+            return sqlite3_column_int64(m_stmt, idx);
         case SQLITE_FLOAT:
-            return sqlite3_column_double(m_stmt, idx);
+            return static_cast<int64_t>(sqlite3_column_double(m_stmt, idx));
         case SQLITE_NULL:
             isNull = true;
             return toWideStr("null");
